[pentair] Made constants and handler packages internal
