+++ addons/binding/org.openhab.binding.pentair/META-INF/MANIFEST.MF
@@ -8,14 +8,12 @@ Bundle-Vendor: openHAB
 Bundle-Version: 2.4.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ClassPath: .
-Export-Package: 
- org.openhab.binding.pentair,
- org.openhab.binding.pentair.handler
 Import-Package: 
  com.google.common.collect,
  gnu.io,
  org.apache.commons.lang,
  org.apache.commons.lang.builder,
+ org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.core.library.types,
  org.eclipse.smarthome.core.thing,
@@ -23,9 +21,6 @@ Import-Package:
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.eclipse.jdt.annotation;resolution:=optional,
- org.openhab.binding.pentair,
- org.openhab.binding.pentair.handler,
  org.osgi.framework,
  org.slf4j,
  org.slf4j.helpers
