+++ addons/binding/org.openhab.binding.pentair/src/main/java/org/openhab/binding/pentair/internal/PentairHandlerFactory.java
@@ -8,7 +8,7 @@
  */
 package org.openhab.binding.pentair.internal;
 
-import static org.openhab.binding.pentair.PentairBindingConstants.*;
+import static org.openhab.binding.pentair.internal.PentairBindingConstants.*;
 
 import org.eclipse.smarthome.core.thing.Bridge;
 import org.eclipse.smarthome.core.thing.Thing;
@@ -16,11 +16,11 @@ import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
-import org.openhab.binding.pentair.handler.PentairEasyTouchHandler;
-import org.openhab.binding.pentair.handler.PentairIPBridgeHandler;
-import org.openhab.binding.pentair.handler.PentairIntelliChlorHandler;
-import org.openhab.binding.pentair.handler.PentairIntelliFloHandler;
-import org.openhab.binding.pentair.handler.PentairSerialBridgeHandler;
+import org.openhab.binding.pentair.internal.handler.PentairEasyTouchHandler;
+import org.openhab.binding.pentair.internal.handler.PentairIPBridgeHandler;
+import org.openhab.binding.pentair.internal.handler.PentairIntelliChlorHandler;
+import org.openhab.binding.pentair.internal.handler.PentairIntelliFloHandler;
+import org.openhab.binding.pentair.internal.handler.PentairSerialBridgeHandler;
 import org.osgi.service.component.annotations.Component;
 
 /**
