+++ addons/binding/org.openhab.binding.pentair/src/main/java/org/openhab/binding/pentair/internal/config/PentairSerialBridgeConfig.java
@@ -6,27 +6,24 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.pentair.config.internal;
+package org.openhab.binding.pentair.internal.config;
 
 import org.apache.commons.lang.builder.ToStringBuilder;
 
 /**
- * Configuration parameters for IP Bridge
+ * Configuration parameters for Serial Bridge
  *
  * @author Jeff James - initial contribution
  *
  */
-public class PentairIPBridgeConfig {
-    /** IP address of destination */
-    public String address;
-    /** Port of destination */
-    public Integer port;
-
+public class PentairSerialBridgeConfig {
+    /** path or name of serial port, usually /dev/ttyUSB0 format for linux/mac, COM1 for windows */
+    public String serialPort;
     /** ID to use when sending commands on the Pentair RS485 bus. */
     public Integer id;
 
     @Override
     public String toString() {
-        return new ToStringBuilder(this).append("address", address).append("port", port).append("id", id).toString();
+        return new ToStringBuilder(this).append("serialPort", serialPort).append("id", id).toString();
     }
 }
